+++ src/server/devices/button/buttonPosition.ts
@@ -1,0 +1,10 @@
+export enum ButtonPosition {
+  topLeft = 1,
+  topRight = 2,
+  midLeft = 3,
+  midRight = 4,
+  botLeft = 5,
+  botRight = 6,
+  top = 11,
+  bottom,
+}