+++ src/server/devices/button/index.ts
@@ -1,3 +1,4 @@
 export * from './button';
 export * from './buttonCapabilities';
 export * from './buttonPressType';
+export * from './buttonPosition';